parts edit

+++ database/migrations/2023_06_07_112015_create_parts_lists_table.php
@@ -22,10 +22,10 @@ return new class extends Migration
             $table->string('lead_time')->nullable();
             $table->string('weight')->nullable();
             $table->string('image')->nullable();
-            $table->integer('surplus_price')->nullable();
-            $table->integer('fs_price')->nullable();
-            $table->integer('navister_price')->nullable();
-            $table->integer('declared_price')->nullable();
+            $table->float('surplus_price')->nullable();
+            $table->float('fs_price')->nullable();
+            $table->float('navister_price')->nullable();
+            $table->float('declared_price')->nullable();
 
 
             $table->timestamps();
