+++ database/migrations/2023_06_13_111705_create_purchase_order_items_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->integer('purchase_order_id');
             $table->integer('item_id');
             $table->integer('qty');
-            $table->integer('price')->nullable();
-            $table->integer('total_price')->nullable();
+            $table->float('price')->nullable();
+            $table->float('total_price')->nullable();
             $table->timestamps();
 
             // $table->foreign('item_id')->references('id')->on('parts_lists')->onDelete('cascade');
