+++ database/migrations/2023_07_18_125751_create_mens_purchase_orders_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('tender_no')->nullable();
             $table->string('wo_no')->nullable();
             $table->date('po_date');
-            $table->integer('total_declared_price_no')->nullable();
+            $table->float('total_declared_price_no')->nullable();
             $table->timestamps();
         });
     }
